Introduce a guardrail for Terraform upgrades
We don't do Terraform upgrade frequently, and when we do we usually forget
about a small bit of manual work we have to do before merging.

To help us not forget we introduce a new guardrail. This guardrail uses a
comment on a PR to remind us to do the manual step, whenever there is a PR
which changes the `.terraform-version` file.