Enable `make ENV forms_apply_all` to start from the previous failure
Implements a checkpoint system for runs of `forms_apply_all`. This is
particularly useful in situations where you need to run the whole set of
Terraform in the correct order, but are encountering errors part way through.
This setting RESUME_FROM_CHECKPOINT=true allows you to skip the previously
successful roots.