ALB response time alert should ignore missing data
We have an alert for the 95th percentile of response times through the ALB being above a second for 5 minutes.

This works nicely in production, but staging is usually in the
"INSUFFICIENT_DATA" state because it gets so little traffic. When it does
finally get a small amount of traffic (every few hours, from observation) it
triggers the OK action for the alarm (because it transitions from
INSUFFICIENT_DATA to OK). This is particularly annoying in the evenings and
weekends for infra people on out-of-hours support, because we're on the mailing
list that these emails go to in staging, so our phones ping somewhat randomly
with new emails about alerts.

This commit makes CloudWatch ignore missing data, so we don't get that happening.