Swap info and warn severity alerts
After discussion, we believe that we had our info and warn severity alerts the
wrong way around. All of the alerts that previously had the warn severity were
actually things we could handle at our leisure - ergo they were informational.
Conversely, all the alerts with the info severity were things we need to act on
soon.

This goes hand in hand with the previous commit which swaps info and warn
destinations, so that warnings go to Slack where we'll see them quickly.