Update forms cli data api to reflect that we now have multiple clusters per environment
Previously we had only one cluster in each environment, and the data api grabbed the arn of that first cluster.
Now that each environment has multiple clusters, this no longer works.
I have changed the forms cli data api to require an additional argument of 'cluster' when invoking - this is the cluster that the db exists in.