Enable sending cloudwatch metrics for apps on dev
Enable runner sending cloudwatch metrics and admin reading them on
dev.

We're going to start sending more CloudWatch metrics from forms-runner
so we can monitor the health of our background jobs that use Solid
Queue. It makes sense to use the same environment variables to enable/
disable sending these metrics, and it would be useful to be able to
test this in dev.

Initially this wasn't enabled for dev as a precaution in case we'd
rack up large costs. As metrics are charged per metric and prorated
per hour and only metered when the metric is being sent the cost would
only become an issue if we were running regular performance tests
using a large number of forms.