Allow forms-runner to send metrics with new namespace
We're renaming our existing metrics to follow the AWS metric naming
conventions, and as part of this we are sending them to a different
namespace. Allow forms-runner to send metrics to the new "Forms"
namesapce.

We're continuing to send the old metrics for now, until we've been
sending them for at least 8 days so we can switch forms-admin to use
the new metrics to display form statistics.