container_images_spec is no longer dependent on Terraform
Previously the tests were dependent on Terraform becuase they invoked it to
find all of the known Terraform functions, so that they could exclude any
strings that started with a function name.

Instead of doing that, we now make the assumption that if the string doesn't
contain a ":" then it isn't a valid image name and we therefore can't test it
for multiple tags.