Use Checkov Docker container instead of installing it locally
If we install it locally, we need to install Python and all the dependencies
first. The container image already has all those things.

We're not using the official bridgecrewio/checkov-action action because its
versions aren't aligned with the version of Checkov, and it makes it hard to
know what version is being used. That action invokes the container anyway, so
we're cutting out the middleman.