Reorder infra-ci workflow steps for faster failures
Previously all of the slow actions (like Checkov) were being done early on,
while the quick and commonly failing tasks were done at the end. This was the
inverse of how it should be.

We have switched them around so that we get faster failures and a shorter
feedback loop.