Change DKIM policy to quarantine, for 25% of failing DMARC checks
Following the advice of NCSC on the topic[1], we've seen that 100% of messages
in the last 28 days passed DMARC checks (and of those that didn't, they failed
because of transient DNS issues), and we're now changing our DMARC policy to
begin quarantining 25% of emails that fail the checks.

In the future we wil ramp this up to 100%, once we're confident we're not
preventing delivery of any emails.

[1] https://www.ncsc.gov.uk/collection/email-security-and-anti-spoofing/mark-spoof-emails-as-spam#section_1