Stop referring to alarms by string in Route53 health checks
In forms/health we configure some Route53 healthchecks. Each healthcheck refers
to a CloudWatch alarm. Previously, the reference was a constructed string which
hapened to match the name of alarm. This prevented Terraform from understanding
the dependencies between resources, and caused race condiitons when creating
them both for the first time.

Instead we should use an in-language reference, so that Terraform can
understand the dependencies.