Increase SQS dead letter queue message retention period
Increase message retention period for the dead letter queues for the
SQS queues that subscribe to the SES notifications SNS topics to 14
days. This matches the retention period for the main SQS queues. This
is to ensure we have enough time to respond to messages on the dead
letter queue before they are deleted.