Increase forms-runner memory and CPU on dev and staging
Since we started running Solid Queue as part of the forms-runner ECS
task, we have seen that we are constantly running at close to 100%
memory, with spikes in CPU utilisation up to 100% in dev and staging.

Increase the CPU and memory so we shouldn't hit these limits. Ideally
we should run Solid Queue in a separate ECS task - if we do this we
might be able to reduce these values.