Replace data lookup in SES alerts with output from a state file
We prefer to explicitly use the outputs from the state file of another root
over using a data lookup. Data lookups create implicit dependencies. State file
outputs are explicit.

This commit exports the queue name from within the ses module and form/ses
root, and imports it in the forms/health root. It also takes the opportunity to
rename some files for clarity.