Temporarily rename dynamodb_table
We'll want to delete the dynamodb tables after we have enabled native state
locking. For now we are renaming them with a default value so that they do not
depend on values we have removed.

This PR makes sure that the policies are still allowed to modify the dynamodb
tables