Allow support role to release the lock on state files
We need the support role to be able to `terraform apply`, therefore it needs to
be able to get/release the lock on state files.

The role also needs to be able to put objects in and get objects from the state
file buckets, but we already gave it those permissions as part of allowing it to
manage_maintenance_page.