Increase period for submission time to send alarm
Increase the time period to 30 minutes. We don't need to react
immediately if submissions are taking longer than expected to send and
it may occur on the weekend or outside working hours when nobody is
available to respond.

We also only care if the time to send it above a certain threshold
for a longer period of time, and if it recovers we don't need to know
about it.