Change submission time to send metric to percentile
Change the alarm to fire if the p95 > 1 minute rather than the
average. We don't care if a few submissions are taking a long time to
send as an anomaly, but we want to know if most submissions are
taking longer than a minute to sned as this indicates we are queuing
submissions faster than we can process them.