Run TFLint on each root individually
Previously we pointed TFLint at the "infra/deployments/forms" directory (for
example) and used the "--recursive" flag to have it check each directory. This
was causing it to evaluate the submodules of those roots as their own roots.

Instead of doing that, we now use the existing lists of roots within the
Makefile to run TFLint against each root individually, and do away with
"--recursive".