Configure the source credentials for GitHub Actions runner in Terraform
Previously it wasn't possible to configure an AWS CodeBuild project as GitHub
Actions runner excluisvely in Terraform becuase of some missing configuration
options. We worked around this using a shell script run after Terraform was
applied.

Since then, the configurations options have been made available in the AWS
Terraform provider and we're now able to retire the shell script and perform
the configuration entirely in Terraform.