Inject Docker secret names into GitHub Actions runner
In a previous commit we began using the parameter ARN as the value of the
environment variable because CodeBuild appeared to support this, by way of not
rejecting it and successfully starting the CodeBuild project.

It turns out it DOESN'T support this. It accepts the ARN, and silently doesn't
create the environment variable if it can't find a parameter with that name
(which it can't, because name != ARN)