Add first secret to Secrets Manager
We are using our own KMS key rather than a managed one because we would like to
have a different key for secrets related to each environment.

This enables administration for admins of the deploy account only. In future we
will add limited power to other roles, and we'll allow the environment accounts
to access the secrets relevant to them