Mark the root filesystem of all apps as readonly
A previous commit made forms-product-page's root filesystem readonly as a
pilot. That was successful and we believe we can now roll this out to all apps.

This commit hasn't removed the configurability in case we need to roll it back
on some but not all apps. When we're confident it works, we'll remove the
configuration option.