Define the variables for different secret types
We have two types of external secrets (used to communicate with external
systems) in the deploy account:
1. environment type secrets - these have the same value in all instances of a
type of environment. But they are not necessarily used in all types. So we need
a list that tells us which environment type needs a given secret. This is the
variable "secrets_in_environment_type"
2. global secrets