Populate secrets with dummy values and expose ids
When creating a new environment we don't know the value of most secrets, so we
want to populate them with a dummy value by default. There are some cases where
we want to set the value, but we will do that in another commit.

We also expose the ids of the secrets as outputs, which allows us to use them in
other roots. This means we can refer to the secrets using an ephemeral resource
later on, and access the secret without showing its value in the state file. e.g.

ephemeral "aws_secretsmanager_secret_version" "databases_forms_admin_app_url_secret_id" {
  secret_id = data.terraform_remote_state.forms_secrets.outputs.internal_secrets_ids["databases_forms_admin_app_url"]
}