TMP:Add the secrets root to the terraform-apply pipeline
I'm not sure this is the right place. We want it to apply automatically, but we
also want some of the secrets to be available for the 'forms/account' root.
We might need to move those secrets to be created in the 'forms/account' root,
keep them as paramters or do some other magic I haven't thought of

Also give the deployer role permissions to manage the secrets