Temporarily remove dependabot package managers
Remove all but the GitHub Actions package manager that dependabot
manages. This is to reduce the overall number of dependencies to try to
diagnose why dependabot is failing to raise PRs.

I've also made the config as "vanilla" as possible - using the default
PR limit and restoring to a daily schedule, just to reduce the
deviation from defaults as much as possible.