Add CloudFront cache policy for /assets
Currently we don't compress or cache any responses from our web apps.

Our Javascript, CSS and fonts are served by our Rails apps.

Serving assets from our Rails apps is simple and makes deployments
easy but it's not ideal for performance.

We could do two things which would speed up the time our pages take to
load and reduce the number of requests our applications have to serve:
- cache assets and serve them from our CDN, Cloudfront
- compress assets to reduce the size of files sent to clients

This commit adds a cache policy to our Cloudfront configuration to serve
the files in the `/assets` directory with compression and caching.

Cloudfront
==========

AWS Cloudfront is the CDN we use to serve our sites. It sits in front of
of the application load balancer (ALB) which routes requests to the
application servers.

This routing is based on the `Host` header. For example if the Host is
`submit.forms.service.gov.uk` then the request is routed to the
runner rails code.

There is one Cloudfront distribution for each environment. All changes
to the configuration affect all applications, Product pages, Admin,
Runner and API.

Currently, Cloudfront:
- passes traffic through the Web Application Firewall (WAF).
- forwards all requests with their headers and cookies to the ALB so
  requests can be served by our application servers.
- serves the error page when the service is unavailable.

What to cache
===

The new cache policy overrides the default for all files in the
`/assets` directory.

These files are the frontend assets generated by Vite, the build tool,
when the docker container is built. They include the javascript, CSS and
fonts used in the design system.

This directory has been chosen because:
- it's always generated by Vite, when the docker container is built.
  Vite adds a hash to the filename based on the contents of the file.
  All references to these files must use the helpers provided by
  vite-rails so we never reference the files directly.
- the files never 'change'. When we make changes to our frontend build,
  the files are rebuilt and the filename changes
- the files are the same for all users. There is no chance of revealing
  personal information or displaying stale data.
- the files are large, so we should get the most benefit from compression
  and caching.

Because the Cloudfront distribution is used by all applications, even
the API would have responses served from /assets cached. Obviously the
API doesn't use /assets but it's worth remembering.

We could apply the cache policy to the whole `/public` directory. The
public directory contains files which have conventional names and could
be linked to from other applications. For example, the `favicon.ico`
file is served from `/public/favicon.ico`.

This doesn't prevent us caching or compressing them but restricting
ourselves to just `/assets` means we can be more confident that we're
not breaking anything.

Origin request policy
===

To create a new cache rule for assets, we need an origin request policy
if we don't want to use the default.

This defines how requests are handled when they are sent to the
application servers through he ALB.

We can change how headers, cookies and query strings are handled.

The default cache policy we use passes all query strings, headers and cookies to the applications.

We use [a managed origin request policy](https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/using-managed-origin-request-policies.html#managed-origin-request-policy-all-viewer)
which does the same thing.

We could be more restrictive and only pass the `Host` header to the
application servers. Assets don't depend on cookies or query strings.

In order to make this change as small as possible though, this commit
sticks to the most conservative choice and matches our default policy.

I think this makes the behaviour assets consistent with the other server
responses and so should be less surprising to developers making changes
to app configuration.

Cache policy
===

The CloudFront cache policy controls how responses are cached. It
controls how long responses are cached and what's used as the cache key.

This commit uses the [CachingOptimized](https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/using-managed-cache-policies.html#managed-cache-caching-optimized) cache policy.

This policy doesn't include headers, cookies or query strings in the cache key.
This is what we want for assets which not depend on any of these.

The biggest impact for us is the `Host` header is not included in the cache key.

This means that files with the same name across our projects will be
cached once and served from the cache. Because we are only serving the
contents of `/assets`, where all files are named with a hash of the
contents we can be confident that the files will be the same.

https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Introduction.html