Move kms creation to the `account` root
When we run the pipeline the deployer role cannot give itself permissions to use
the kms key. Instead we should move the responsibility to give access to kms
keys to the `account` root (which we apply manually from a dev laptop)

This means when we create a new environment, we will apply the `account` root
manually, and then the pipelines can be applied automatically without having to
stop to give permissions to the deployer role.