Use the alphagov version of this runner
Ubuntu ARM64 hosted runners are currently in public preview, accessible through
the ubuntu-24.04-arm and ubuntu-22.04-arm preset labels. They are not available
in private repos.

The runners we intended to use are in Alphagov (not GitHub runners), but happen
to have the same labels as GitHub runners. This means our action started failing
and we didn't know why.

We have asked Alphagov to rename this runner so that we can tell them apart.