Allow large POSTs for specific endpoints
Previously, we loosened the WAF rules in the past to allow for large
POSTs when uploading files, and for bulk options uploads. This was
achieved by disabling the managed `SizeRestrictions_BODY` rule.

Instead, we now have two targetted rules that allow large requests to
the endpoints that need it, reinstating the default
`SizeRestrictions_BODY` rule.

All rules are set to `count` action for now, so we can monitor for
false positives after applying, and spot any other endpoints that
might need to be added. Once we're happy that the new rules are
working correctly, we can change the action to `allow` and remove the
`SizeRestrictions_BODY` `rule_action_override`.

File upload is allowed when the request has a `multipart/form-data`
content type and the URI path matches a regex for form preview or
submission. The maximum file size is set to 100MB by default. This is
large enough to allow users to mistakenly upload large files and get
the app-level error handling to warn them, rather than hitting the WAF
and getting a generic error.

In future, we should consider modifying the application to read the
Content-Length header and error if it exceeds a certain size. This will
allow us to set the limit in the WAF to the app limit + a small
buffer - if the Content-Length header has been modified by a user, we
can fairly safely assume that they are trying to be malicious, in which
case a WAF error is appropriate.

Multi-option bulk uploads are allowed when the request URI path matches
a regex for the bulk options upload endpoints. The maximum size is set
to 106400 bytes (103.91 KB) by default, which is based on allowing
1000 options, 100 characters each, plus some overhead for the rest of
the form data (400 bytes). It's fairly unlikely that anyone will need
to upload more than this, but it can be adjusted if necessary.