Fix bundle installed dependencies for development
When running in RAIL_ENV=development mode, Rails requires the "debug"
gem to be present. However becuase we build the containers from our
Dockerfile for production images, we skip install development
dependencies. This adds the BUNDLE_WITHOUT arg to just ignore the "test"
dependencies, but still include the "development" dependencies (which
includes the "debug" gem"). It does now mean all our development
dependences are installed in our local dev containers, but not the end
of the world.