Increase the max size for form post bodies 7-100kb
We were seeing some COUNT hits on some forms where the body was larger
than the default 7kb limit. Although these are larger than the
server-side validation would allow, we want to allow the user to see
the pretty error message from the form rather than a WAF error.

I've also renamed the resources and changed the order within the rule
group to make it cleaner.