New terraform upgrade script using tfupdate
Rather than the manual ruby script, we can use tfupdate to do the heavy
lifting of finding and updating the versions of terraform and the
providers.

Tfupdate also has the ability to update the lock files for us using an
in-memory cache of the plugins, removing the need for us to `init` and
download the plugins multiple times.

This does switch us away from fuzzy version matching to exact versions,
but that is a good thing as it means we know exactly what version we are
using.

We no longer read the terraform version from terraform_version.tf.json,
instead we read it from .terraform-version. This ensures that the
version used in codebuild and locally are the same, as tfenv reads
this file.