Stop filtering out log messages containing /up
Since Rails 8, the health check controller logs are filtered out from
the log stream by default [[1]]. As all our apps are now using Rails 8
and have `config.silence_healthcheck_path` enabled [[2]], we don't need
to filter out logs with "/up" in our log forwarding configuration.

We could keep it around just in case, but it's also filtering out lines
we actually want, and making the filter more specific has proven to be
quite hard (see previous commits for this line of code), so it's better
to get rid of it.

[1]: https://github.com/rails/rails/pull/52789
[2]: https://guides.rubyonrails.org/configuring.html#config-silence-healthcheck-path