Add a make target to open a tf shell within a root
Sometimes it's useful to be able to use `terraform` commands in the
context of an environment (e.g. to inspect state, or to taint resources)
without having to re-run `terraform init` and set all the required
environment variables manually.

This commit adds a `make tf_shell` target which will open an interactive
shell with the correct working directory and environment variables set.