Get forms/auth0 working properly when disabled.
We previously pinned the provider because auth0 made the settings
not-nullable. This meant that when Auth0 was disabled and the settings
were empty strings, the provider would error out.

Instead, pass dummy values to the provider when Auth0 is disabled. The
credentials are not used because the actual auth0 resources are all
conditionally created based on whether Auth0 is enabled or not.

This may bite us in future if the provider does more validation of
the credentials, but for now it lets us keep the provider up to date.