docker-compose: Remove need for .env file
We were using the .env file for Docker Compose to set the default value
of the environment variable holding the forms-api base URL.

However, it is better not to have a .env file committed in a repo,
because doing so makes it harder for developers to put their own
customizations in that file.

Instead change the compose file to set the default for the environment
variable via interpolation. This does result in a little bit of
duplication, but that won't be a problem for much longer as we're
stopping using forms-api soon anyway.

We also add .env to a .gitignore file for the directory with the
Composefile, so developers can in future add things to that file without
worrying about accidentally committing it.