docker-compose: Use forms-admin to serve form documents to forms-runner
We want to stop using forms-api and start serving form documents from
forms-admin instead, to simplify our system architecture.

This commit updates the Docker Composefile for our local development
environment to configure forms-runner to use forms-admin instead of
forms-api.

We still need the forms-api app, as forms-admin expects to be able to
communicate with it still.

Note that this commit does not change any of our environments in AWS
(including our production environment), this is only for our local
developer setups.