Allow non-admin users read access to Console UXC
AWS has added a feature where the console header can be coloured.

There is no terraform resource for this yet, so the colours have been
set manually:

- Production: red
- Staging: orange (warm warning)
- Dev: green (safe to proceed)
- User-research: purple (distinctive)

Read access to the colours isn't part of the ReadOnlyAccess policy, so
we need to add a custom policy to allow read access to the UX
Customisation service and attach it to our non-admin roles.