Increase resources for forms-admin in production to match forms-api
We are about to change our architecture so that forms-admin serves form
documents to forms-runner, and we'll get rid of forms-api. To make it
less likely that the additional requests to forms-admin cause
performance issues, this commit increases the resource allocation for
forms-admin to match what forms-api has. Specifically, we increase the
vCPU and memory allocation for forms-admin, and allow the autoscaling to
create more instances than previously.

Note that we only need to do this in production, as for the other
environments forms-api has the same resources as forms-admin, and we
haven't seen any issues with that.