BAU: Simplify target group names to reduce length
`user-research` was failing to deploy because the internal target group
name was too long (over 32 characters). Remove the env name from the
target group names to ensure they fit within the limit.

Not changing the external target group name in the same way, as that
would cause downtime during deployment. The internal target group is
currently unused in production so this is a safe operation.

This will require the ECS services to be forced to redeploy to get them
to register with the new target groups.