BAU: group github actions PRs together
This will reduce the number of rebases we will have to do, due to the
'branch must be up to date before merging' setting on this repo.

As we now run the workflow when the workflow file itself is modified,
we can be sure that any version updates will be tested.