Grant rds_extension role in forms-admin database setup
Add the statment to grant the "rds_extension" role to the
"forms-admin-app" in the SQL script to prepare the forms-admin
database.

This is so that if we need to re-create the database, this command is
run so that the migrations have the ability to add postgres
extensions. We need to run this SQL manually in the RDS query editor
to grant this role for now.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>