BAU: only define the list of actions once
Previously, we had two lists of action names:
- in the `getopts` switch/case, where we check if the action is valid
- in the final 'action runner' switch/case, where we run the actions

This has been reduced to one list - in the action runner. If an action
exists and is valid, it is run. If not, we fail over to `usage`.

This does mean that we don't validate the actions until we have valid
values for all the other options, but as this script is only run via the
Make targets, this shouldn't matter.