BAU: implement terraform plugin caching
The standard behaviour of terraform is to download and install the
provider binaries into `.terraform/providers/**` directories. This is
useful if modules have substantially different provider versions within
a project.

In our case, we use the same set of plugin versions across all
deployments. Each of these `providers` directories is 800+ MB, which
adds up quickly with the number of modules in this repo (~20GB).

Using a single plugin cache directory means that we:
1. Only have one single 800MB directory on development laptops
2. No longer have to download the providers again after the first `init`

https://developer.hashicorp.com/terraform/cli/config/config-file#provider-plugin-cache

This plugin cache will grow over time. a `clear-tf-plugin-cache` target
has been added to the Makefile to facilitate easy cleanup (full purge)