Give deployer access to awscc
The deployer needs additional permissions to manage resources via the
Cloud Control API, which is used by the `awscc` Terraform provider.

Details of what this means:

These permissions are required for the `awscc` provider to create,
update, delete, and list Cloud Control API resources, which includes the
Application Signals SLO resources.
See: https://docs.aws.amazon.com/cloudcontrolapi/latest/userguide/security.html

It seems super scary to give `cloudformation:*` permissions, but the
Cloud Control API is implemented as a layer on top of CloudFormation,
and the `awscc` provider needs these permissions to manage resources.
These permissions are essentially saying 'allow the deployer to use the
Cloud Control API to attempt to CRUDL any resource'. The deployer role
will still be limited by the other permissions it has, so it won't be
able to create resources we haven't explicitly given it permission for.
eg. if we removed `s3:CreateBucket` from the deployer role, it wouldn't
be able to create S3 buckets, even though it has
`cloudformation:CreateResource` permission.