Convert the submission delivery SLO to metric_stat
CloudFormation (and therefore awscc provider) does not yet support the
period parameter in the SEARCH function, so we have to use metric_stat
to define the SLOs. We have to hardcode the submission delivery
methods we use (Email and S3) until CloudFormation supports this.
If we add any other submission delivery methods, they must be added to
the variable defined here or the SLOs will not account for them.

An issue has been raised with AWS to track this:
https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/2372