Move Route 53 internal zone to account root
This means that the deployer won't need permission to create the zone,
only to manage the VPC association.

We create a dummy VPC in the account root to allow the internal zone to
be created before the environment is deployed. The deployer will
then reassociate the zone with the real VPC in the environment
deployment. Internal zones can't be created without a VPC, but the vpc
does not exist until `environment` is deployed.

Empty VPCs are not billable, so this does not incur any cost, just a
bit of weirdness.