Create and use `access-logs-bucket` module
We created access log buckets in two ways:
1. For bucket access logs, we created the resources piecemeal for each
   bucket type (secure-bucket, public-bucket, state-bucket)
2. For other access logs, we used the existing `secure-bucket` module
   and reconfigured it.

This allows us to create one specific configuration for how
we want access log buckets to be configured.

We now have a lifecycle rule for access log buckets to move objects to
STANDARD_IA after 30 days, GLACIER_IR after 90 days, and expire after
365 days. We also no longer have versioning enabled on access log
buckets: a new file is uploaded each time, so versioning is not helpful,
and it just increases storage costs.