Rename log groups for CRIBL migration
We need to rename log groups to match the convention that the Cyber team has
set. They are migrating from CSLS to CRIBL and this is part of the preparation.
In CRIBL they have setup pipelines that filter logs based on the name of log
groups, so they have given us a convention to follow

This includes a small refactor to make sure we don't forget to change the log
group name for ecs logs where it is used