Add kinesis stream to `deploy` account
Following the Cyber Engineering's guide
(https://github.com/CO-Cyber-Security/cribl-cloudwatch-kinesis/tree/main/logs_account),
this creates a kinesis stream in the deploy account. We also give permission to
all account to subscribe to the stream.

TODO: We have narrowed down the iam policy (cribl_kinesis) to only have access to the
kinesis stream that we create here (instead of "*"). Check with Cyber that this
is ok