Setup subscription filter for ecs logs
We're setting up the subscription filter as we did with csls, in the ecs-service
module. This means we need to pass in the values for the destination arn and the
role to use so there's quite a bit of plumbing in here.

The alternative would be to pass the name of the log group "up" to the root that
calls the module, and create the subscription filter at a higher level. I think
both are a bit cumbersome, but this one follows a more familiar pattern for us.

The solution that cyber provided would create all subscriptions in one go
(https://github.com/CO-Cyber-Security/cribl-cloudwatch-kinesis/blob/3ff761dbe400060d954c915c097aa469e0d9313d/production_account/cloudwatch_subscription.tf#L11),
however we'd have to supply the log group names for all roots. This would mean
creating the subscription _after_ all roots have been applied. Because of the
delay we would have a longer gap between the log and subscription creation. I'm
not too sure if we care about that