Create second destination in us-east-1 to same kinesis stream
We have some logs groups in us-east-1. When we create a subscription, the
subscription filter has to live in the same region, and we cannot set a
destination in a different region.

So we create everything (log group, subscription filter and destination) in
us-east-1, but target the kinesis stream in eu-west-2