Update infracost config post-forms-api
Because `git` only deals with files, if someone else deletes a directory
that was previously tracked, other users will still see that directory
on their local machines. The previous logic in the script did not
account for this, leading to stale entries in the Infracost config.

By adding a check to ensure that directories contain files before
including them in the Infracost configuration, we can prevent
stale entries from persisting.