Restore data_api cli tool
When the API was removed, the data_api command was removed from the CLI tool in this PR:
https://github.com/alphagov/forms-deploy/pull/1769

This broke the CLI tool as it still requires the file when starting up.

I think it was a mistake to remove the command.

This commit restores the data_api command and tests exactly as they
were.