BAU: Create data modules for cribl / csls arns
CSLS:

Rather than hardcoding these arns in multiple places, store them in
a data module that can be referenced where needed.

CRIBL:

We're going to create the SQS queue / IAM role in the deployer account.
Instead of having to use `remote-state` to get these values later,
we can
- specify the names of the resources here, and then create them using
  these values
- Generate the resulting arns here (as we know what they will be), and
  export them via outputs, so they can be referenced later by importing
  this module.

It does mean that we need the deployer account to be fully deployed
first, but that's a fairly safe assumption to make.