Allow log shipping to Cribl from S3 via SQS
We have removed external log shipping module, because it deals with
bucket object ownership controls internally. It's much easier to handle
that within this module directly, as we will need that for cribl too. If
we had both modules, we'd have to carefully coordinate ownership control
creation and deletion when switching between the two log shippers.