BAU: Fix `file-upload` S3 access logging
Past mistakes meant that S3 access logs were not being sent to Cyber as
intended, as the bucket notification was being switched to using the
modularised access logging. Remove the direct usage of
`cyber_s3_log_shipping` module, and rely on the internal handling of
the `secure-bucket` module to set up the log shipping.