Run Solid Queue worker in forms-runner ECS task
Currently, the forms-runner-queue-worker ECS task is not being
re-deployed with new forms-runner code as part of the deployment
pipeline.

Temporarily switch back on the Solid Queue worker process in the
forms-runner ECS tasks and scale the forms-runner-queue-worker tasks
to 0. We can switch this back when the forms-runner-queue-worker is
being successfully deployed.