BAU: use container_image_uri to retag, not json
Previously, we created the json file to update the container image tag,
and then read that json file to get the current image uri to retag it.

Now, we just pass the container image uri directly as an environment
variable to the CodeBuild project, simplifying the process. This has
the benefit of making the forms-runner pipeline clearer, as it could
read as if we were specifically reading from `forms-runner-image-defs-json`
over `queue-worker-image-defs-json`. In fact, either file would have
worked.